taggor

+++ src/output_h.c
@@ -2,50 +2,50 @@
 
 t_data			*build_hd(void)
 {
-	short			val;
+	int				val;
 	char			*tmp;
 
-	val = va_arg(VARS, short);
+	val = va_arg(VARS, int);
 	tmp = ft_itoa((int)val);
 	return (create_data(tmp, NULL));
 }
 
 t_data			*build_ho(void)
 {
-	short			val;
+	int				val;
 	char			*tmp;
 
-	val = va_arg(VARS, short);
+	val = va_arg(VARS, int);
 	tmp = octal_string((unsigned int)val);
 	return (create_data(tmp, NULL));
 }
 
 t_data			*build_hu(void)
 {
-	short			val;
+	int				val;
 	char			*tmp;
 
-	val = va_arg(VARS, short);
+	val = va_arg(VARS, int);
 	tmp = ft_uitoa((unsigned int)val);
 	return (create_data(tmp, NULL));
 }
 
 t_data			*build_hx(void)
 {
-	short			val;
+	int				val;
 	char			*tmp;
 
-	val = va_arg(VARS, short);
+	val = va_arg(VARS, int);
 	tmp = hexa_string((unsigned int)val, FALSE);
 	return (create_data(tmp, NULL));
 }
 
 t_data			*build_h_maj_x(void)
 {
-	short			val;
+	int				val;
 	char			*tmp;
 
-	val = va_arg(VARS, short);
+	val = va_arg(VARS, int);
 	tmp = hexa_string((unsigned int)val, FALSE);
 	return (create_data(tmp, NULL));
 }
